ReactOS 0.4.0 for CLT2016 was already released! Tag the branch.

+++ reactos/dll/win32/mapi32/imalloc.c
@@ -96,9 +96,9 @@ static ULONG WINAPI IMAPIMalloc_fnRelease(LPMALLOC iface)
 /**************************************************************************
  * IMAPIMalloc_Alloc
  */
-static LPVOID WINAPI IMAPIMalloc_fnAlloc(LPMALLOC iface, SIZE_T cb)
+static LPVOID WINAPI IMAPIMalloc_fnAlloc(LPMALLOC iface, DWORD cb)
 {
-    TRACE("(%p)->(%ld)\n", iface, cb);
+    TRACE("(%p)->(%d)\n", iface, cb);
 
     return LocalAlloc(LMEM_FIXED, cb);
 }
@@ -106,9 +106,9 @@ static LPVOID WINAPI IMAPIMalloc_fnAlloc(LPMALLOC iface, SIZE_T cb)
 /**************************************************************************
  * IMAPIMalloc_Realloc
  */
-static LPVOID WINAPI IMAPIMalloc_fnRealloc(LPMALLOC iface, LPVOID pv, SIZE_T cb)
+static LPVOID WINAPI IMAPIMalloc_fnRealloc(LPMALLOC iface, LPVOID pv, DWORD cb)
 {
-    TRACE("(%p)->(%p, %ld)\n", iface, pv, cb);
+    TRACE("(%p)->(%p, %d)\n", iface, pv, cb);
 
     if (!pv)
         return LocalAlloc(LMEM_FIXED, cb);
@@ -132,7 +132,7 @@ static void WINAPI IMAPIMalloc_fnFree(LPMALLOC iface, LPVOID pv)
 /**************************************************************************
  * IMAPIMalloc_GetSize
  */
-static SIZE_T WINAPI IMAPIMalloc_fnGetSize(LPMALLOC iface, LPVOID pv)
+static DWORD WINAPI IMAPIMalloc_fnGetSize(LPMALLOC iface, LPVOID pv)
 {
     TRACE("(%p)->(%p)\n", iface, pv);
     return LocalSize(pv);
